# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PixEz Sync 后端同步服务 API。除 /mirror/** Pixiv 形态镜像接口外，系统 API 均返回 {success,message,data} 标准响应。",
        "title": "PixEz Sync Server API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/pixez/illusts/{illust_id}/mirror": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Returns the latest mirror task status for a Pixiv illustration.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mirror"
                ],
                "summary": "Check illustration mirror status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pixiv illustration ID",
                        "name": "illust_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.MirrorStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Enqueues an async task to mirror a Pixiv illustration detail response and its original images.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mirror"
                ],
                "summary": "Enqueue illustration mirror task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pixiv illustration ID",
                        "name": "illust_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.MirrorStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/mirror/v1/illust/detail": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Reads a cached Pixiv /v1/illust/detail response and rewrites pximg URLs to this server's /mirror/pximg path. This endpoint intentionally returns the Pixiv response shape, not the standard PixEz Sync API envelope.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pixiv Mirror"
                ],
                "summary": "Get mirrored Pixiv illustration detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pixiv illustration ID",
                        "name": "illust_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.BasicErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.BasicErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.BasicErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "model.BasicErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "model.MirrorStatusResponse": {
            "type": "object",
            "properties": {
                "failed_count": {
                    "type": "integer"
                },
                "illust_id": {
                    "type": "integer"
                },
                "mirrored": {
                    "type": "boolean"
                },
                "request_urls_json": {
                    "type": "string"
                },
                "retry_urls_json": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "success_count": {
                    "type": "integer"
                },
                "task_id": {
                    "type": "string"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}